+++ CHANGELOG.md
@@ -12,10 +12,28 @@ Installation
 Pragmas and options
 -------------------
 
+* When the new pragma option `--opaque` is used, then every function
+  definition, mutual block, `unquoteDecl` or `unquoteDef` that is not
+  explicitly marked as opaque or transparent is put into a separate
+  opaque block ([#6934](https://github.com/agda/agda/issues/6934)).
+
+  Local definitions in where clauses are put into the same opaque
+  block as their parent definitions, unless `opaque` or `transparent`
+  are used.
+
+  Note that definitions in let expressions or before a record type's
+  last field declaration are not made opaque.
+
 Syntax
 ------
 
 Additions to the Agda syntax.
+
+* [**Breaking**] The new keyword `transparent` introduces a block of
+  transparent definitions
+  ([#6934](https://github.com/agda/agda/issues/6934)). This keyword
+  can be used to make definitions inside opaque blocks transparent,
+  and also to make definitions transparent when `--opaque` is used.
 
 Language
 --------